main

+++ termux-manager.sh
@@ -617,46 +617,51 @@ login() {
 tli_wrapper() {
     DEINE_WEBSITE_URL="https://term-master.netlify.app/pkg/"
 
-COMMAND=$1
-PACKAGE_NAME=$2
+    COMMAND=$1
+    PACKAGE_NAME=$2
 
-# Vollständige URL der ZIP-Datei
-ZIP_URL="${DEINE_WEBSITE_URL}/${PACKAGE_NAME}.zip"
+    # Vollständige URL der ZIP-Datei
+    ZIP_URL="${DEINE_WEBSITE_URL}/${PACKAGE_NAME}.zip"
 
-# Funktion, um die ZIP-Datei herunterzuladen
-download_zip() {
-    echo "Download..."
-    wget -q --spider $ZIP_URL
+    # Funktion, um die ZIP-Datei herunterzuladen
+    download_zip() {
+        echo "Versuche, ZIP-Datei herunterzuladen..."
+        wget -q --spider $ZIP_URL
 
-    if [ $? -eq 0 ]; then
-        echo "Download ZIP"
-        wget $ZIP_URL
-        mv $PACKAGE_NAME.zip .cache/tli
-        pkg install zip -y
-        cd .cache/tli
-        mkdir $PACKAGE_NAME
-        mv $PACKAGE_NAME.zip $PACKAGE_NAME/
-        unzip $PACKAGE_NAME.zip 
-        pkg install termux-create-package -y
-        termux-create-package *.json
-        dpkg -i *.deb 
-        cd $HOME
-    else
-        echo "Keine ZIP-Datei gefunden. Führe pkg aus..."
-        pkg $COMMAND $PACKAGE_NAME
-    fi
-}
+        if [ $? -eq 0 ]; then
+            echo "ZIP-Datei gefunden. Herunterladen..."
+            wget $ZIP_URL -O "${PACKAGE_NAME}.zip"
+            echo "Entpacke ZIP-Datei..."
+            mkdir ${PACKAGE_NAME}
+            unzip "${PACKAGE_NAME}.zip" -d "${PACKAGE_NAME}"
 
-# Hauptlogik des Skripts
-case $COMMAND in
-    install)
-        download_zip
-        ;;
-    *)
-        echo "Führe pkg mit originalen Parametern aus..."
-        pkg "$@"
-        ;;
-esac
+            # Prüfen, ob die JSON-Datei existiert
+            if [ -f "${PACKAGE_NAME}/${PACKAGE_NAME}.json" ]; then
+                echo "Erstelle Debian-Paket aus JSON..."
+                pkg install termux-create-package -y
+                termux-create-package "${PACKAGE_NAME}/${PACKAGE_NAME}.json" -o "${PACKAGE_NAME}*.deb"
+
+                echo "Installiere Debian-Paket..."
+                dpkg -i "${PACKAGE_NAME}/${PACKAGE_NAME}*.deb"
+            else
+                echo "JSON-Datei für das Paket wurde nicht gefunden."
+            fi
+        else
+            echo "Keine ZIP-Datei gefunden. Führe pkg aus..."
+            pkg $COMMAND $PACKAGE_NAME
+        fi
+    }
+
+    # Hauptlogik des Skripts
+    case $COMMAND in
+        install)
+            download_zip
+            ;;
+        *)
+            echo "Führe pkg mit originalen Parametern aus..."
+            pkg "$@"
+            ;;
+    esac
 }
 
 help() {
